Added Hilt annotations + package "presentation" renamed
