+++ app/build.gradle
@@ -108,3 +108,7 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
+
+kapt {
+    correctErrorTypes true
+}
