+++ app/src/main/java/dev/tutushkin/allmovies/AllMoviesApplication.kt
@@ -1,0 +1,21 @@
+package dev.tutushkin.allmovies
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+// TODO Add loader
+// TODO Add multi language support
+// TODO Add save favorites
+// TODO Add movie search
+// TODO Add info about actors (new screen)
+// TODO Use Navigation
+// TODO Use DI
+// TODO Add column alignment to the RecyclerView
+// TODO Optimize image sizes dynamically based on a display/network speed/settings
+// TODO Add tests
+// TODO Add logging
+// TODO Replace Toasts with SnackBars
+// TODO Add trailers
+
+@HiltAndroidApp
+class AllMoviesApplication : Application()
