+++ app/src/main/java/dev/tutushkin/allmovies/ui/MainActivity.kt
@@ -1,0 +1,15 @@
+package dev.tutushkin.allmovies.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+import dev.tutushkin.allmovies.R
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
