+++ app/src/main/java/dev/tutushkin/allmovies/ui/moviedetails/viewmodel/MovieDetailsViewModel.kt
@@ -1,14 +1,17 @@
-package dev.tutushkin.allmovies.presentation.moviedetails.viewmodel
+package dev.tutushkin.allmovies.ui.moviedetails.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.tutushkin.allmovies.BuildConfig
 import dev.tutushkin.allmovies.domain.movies.MoviesRepository
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class MovieDetailsViewModel(
+@HiltViewModel
+class MovieDetailsViewModel @Inject constructor(
     private val moviesRepository: MoviesRepository,
     private val id: Int
 ) : ViewModel() {
