+++ app/src/main/java/dev/tutushkin/allmovies/ui/moviedetails/viewmodel/MovieDetailsViewModelFactory.kt
@@ -1,16 +1,17 @@
-package dev.tutushkin.allmovies.presentation.movies.viewmodel
+package dev.tutushkin.allmovies.ui.moviedetails.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import dev.tutushkin.allmovies.domain.movies.MoviesRepository
 
-class MoviesViewModelFactory(
-    private val repository: MoviesRepository
+class MovieDetailsViewModelFactory(
+    private val repository: MoviesRepository,
+    private val id: Int
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MoviesViewModel::class.java)) {
-            return MoviesViewModel(repository) as T
+        if (modelClass.isAssignableFrom(MovieDetailsViewModel::class.java)) {
+            return MovieDetailsViewModel(repository, id) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
