+++ app/src/main/java/dev/tutushkin/allmovies/ui/movies/view/MovieViewHolder.kt
@@ -1,4 +1,4 @@
-package dev.tutushkin.allmovies.presentation.movies.view
+package dev.tutushkin.allmovies.ui.movies.view
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
