+++ app/src/main/java/dev/tutushkin/allmovies/ui/movies/viewmodel/MoviesViewModelFactory.kt
@@ -1,17 +1,16 @@
-package dev.tutushkin.allmovies.presentation.moviedetails.viewmodel
+package dev.tutushkin.allmovies.ui.movies.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import dev.tutushkin.allmovies.domain.movies.MoviesRepository
 
-class MovieDetailsViewModelFactory(
-    private val repository: MoviesRepository,
-    private val id: Int
+class MoviesViewModelFactory(
+    private val repository: MoviesRepository
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MovieDetailsViewModel::class.java)) {
-            return MovieDetailsViewModel(repository, id) as T
+        if (modelClass.isAssignableFrom(MoviesViewModel::class.java)) {
+            return MoviesViewModel(repository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
