+++ app/src/main/res/layout/fragment_movies_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".presentation.MainActivity">
+    tools:context=".ui.MainActivity">
 
     <ImageView
         android:id="@+id/movies_list_target_image"
