+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/movieDetailsFragment"
-        android:name="dev.tutushkin.allmovies.presentation.moviedetails.view.MovieDetailsFragment"
+        android:name="dev.tutushkin.allmovies.ui.moviedetails.view.MovieDetailsFragment"
         android:label="MovieDetailsFragment">
         <argument
             android:name="movieId"
@@ -14,7 +14,7 @@
     </fragment>
     <fragment
         android:id="@+id/moviesFragment"
-        android:name="dev.tutushkin.allmovies.presentation.movies.view.MoviesFragment"
+        android:name="dev.tutushkin.allmovies.ui.movies.view.MoviesFragment"
         android:label="MoviesFragment">
         <action
             android:id="@+id/action_moviesFragment_to_movieDetailsFragment"
